requirements.txt
# Flask web framework
Flask==3.0.2

gunicorn==21.2.0

google-cloud-storage==2.9.0

google-cloud-aiplatform
google-generativeai>=0.3.0

llama-index>=0.6.0
llama-index-multi-modal-llms-gemini
llama-index-embeddings-gemini
llama-index-llms-gemini
